📱 fix: homepage performance analytics responsiveness
Changed homepage performance analytics section on smaller devices to only show a button to the dashboard due to issues with recharts rendering.

Abstracted rounds table to its own component
